refactor: remove unnecessary folder
