+++ src/components/working-experience-component/Fila.tsx
@@ -2,13 +2,13 @@
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 // styled-components
-import { BrownCenterTitle } from "../../styled-component/Title";
+import { BrownCenterTitle } from "../styled-component/Title";
 
-export default function PoppingDanceInstructor() {
+export default function Fila() {
   return (
     <>
       <Card
-        id="popping-dance-instructor-container"
+        id="fila-container"
         sx={{
           background: "white",
           border: "5px solid rgba(3, 102, 214, 0.3)",
@@ -29,19 +29,13 @@ export default function PoppingDanceInstructor() {
               textAlign: "left",
             }}
           >
-            Popping Dance Instructor - from July 2023
+            FILA - from April 2022
           </BrownCenterTitle>
-          <span style={{ flex: 1, fontWeight: 700 }}>
-            RMIT University Dacning Club "Funkdelics" · Freelance
-          </span>
-          <span style={{ flex: 1 }}>
-            This is my second time to be the instructor for popping dance in
-            this club. I delivered my knowledges and understanding of dancing to
-            help the beginner dancers get used with popping dance.
-          </span>
+          <span style={{ flex: 1, fontWeight: 700 }}>Customer Assistance</span>
           <ul style={{ flex: 1 }}>
-            <li>Teaching Popping Dance in Beginner Level</li>
-            <li>Deliver the relevant knowledges</li>
+            <li>General customer service including cashier and consulting</li>
+            <li>Bilingual customer service</li>
+            <li>Replenish the products</li>
           </ul>
         </CardContent>
       </Card>
